+++ __tests__/nthEquals.test.js
@@ -3,7 +3,7 @@ import { nthEquals } from '../src/nthEquals'
 it('checks if an element of array deeply equals value', () => {
   const arr = [{ a: 1 }]
 
-  expect(nthEquals({ a: 1 }, 0, arr)).toBe(true)
-  expect(nthEquals({ a: 2 }, 0, arr)).toBe(false)
-  expect(nthEquals({ a: 1 }, 1, arr)).toBe(false)
+  expect(nthEquals(0, { a: 1 }, arr)).toBe(true)
+  expect(nthEquals(0, { a: 2 }, arr)).toBe(false)
+  expect(nthEquals(1, { a: 1 }, arr)).toBe(false)
 })