+++ __tests__/propEquals.test.js
@@ -3,7 +3,7 @@ import { propEquals } from '../src/propEquals'
 it('checks if a property deeply equals value', () => {
   const obj = { a: [1] }
 
-  expect(propEquals([1], 'a', obj)).toBe(true)
-  expect(propEquals([2], 'a', obj)).toBe(false)
-  expect(propEquals(undefined, 'b', obj)).toBe(false)
+  expect(propEquals('a', [1], obj)).toBe(true)
+  expect(propEquals('a', [2], obj)).toBe(false)
+  expect(propEquals('b', undefined, obj)).toBe(false)
 })